chg: refactor statesync fetcher
